requirements.txt
PyYAML~=6.0.1
PyYAML~=6.0.1
requests_html~=0.10.0
